python_encryption/requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
reportlab
cryptography==39.0.1
pycryptodome==3.17